+++ services/workflows-service/src/workflow/hook-callback-handler.service.ts
@@ -216,11 +216,11 @@ export class HookCallbackHandlerService {
     // @ts-expect-error - we don't validate `context` is an object
     this.setNestedProperty(context, attributePath, result);
     // @ts-expect-error - we don't validate `context` is an object
-    context.documents = [
+    context.documents =
       // @ts-expect-error - we don't validate `context` is an object
-      ...(context.documents?.filter(document => document.type !== 'identification_document') ?? []),
-      ...persistedDocuments,
-    ];
+      context.documents?.filter(document => document.type !== 'identification_document') ?? [];
+    // @ts-expect-error - we don't validate `context` is an object
+    context.kycDocuments = persistedDocuments;
 
     return context;
   }